chore: split Next plugin internals
