+++ packages/vitest-plugin-rsc/src/nextjs/route-manifest-plugin.ts
@@ -1,0 +1,326 @@
+import { Buffer } from "node:buffer";
+import fs from "node:fs";
+import path from "node:path";
+import type { Plugin } from "vite";
+import { createProjectRequire, getProjectRoot, normalizePath } from "./plugin-utils";
+
+const virtualNextRouteManifestId = "\0vitest-plugin-rsc:next-route-manifest";
+const virtualNextRouteManifestPublicId = "virtual:vitest-plugin-rsc/next-routes";
+const virtualNextRouteTreeIdPrefix = "\0vitest-plugin-rsc:next-route-tree?";
+const virtualNextRouteTreePublicId = "virtual:vitest-plugin-rsc/next-route-tree";
+const virtualNextRouteEmptyModuleId = "\0vitest-plugin-rsc:next-route-empty-module";
+const virtualNextRouteEmptyModulePublicId = "virtual:vitest-plugin-rsc/next-route-empty-module";
+
+type NextRouteManifestBuildEntry = {
+  route: string;
+  appDir: string;
+  appPath: string;
+  appPaths: readonly string[];
+  allNormalizedAppPaths: readonly string[];
+  pageFile: string;
+};
+
+export function useNextRouteManifest(): Plugin {
+  let root = process.cwd();
+
+  return {
+    name: "next-rsc-route-manifest",
+    enforce: "pre",
+    configResolved(config) {
+      root = getProjectRoot(config);
+    },
+    resolveId(source) {
+      const [sourceFile] = source.split("?");
+      if (
+        sourceFile &&
+        path.isAbsolute(sourceFile) &&
+        !fs.existsSync(sourceFile) &&
+        isInNextAppDir(root, sourceFile)
+      ) {
+        return virtualNextRouteEmptyModuleId;
+      }
+
+      if (source === virtualNextRouteManifestPublicId) {
+        return virtualNextRouteManifestId;
+      }
+      if (source.startsWith(`${virtualNextRouteTreePublicId}?`)) {
+        return `${virtualNextRouteTreeIdPrefix}${source.slice(virtualNextRouteTreePublicId.length + 1)}`;
+      }
+      if (source === virtualNextRouteEmptyModulePublicId) {
+        return virtualNextRouteEmptyModuleId;
+      }
+    },
+    async load(id) {
+      if (id === virtualNextRouteEmptyModuleId) {
+        return "export default function VitestNextEmptyRouteModule() { return null; }";
+      }
+
+      if (id.startsWith(virtualNextRouteTreeIdPrefix)) {
+        const params = new URLSearchParams(id.slice(virtualNextRouteTreeIdPrefix.length));
+        const pageFile = params.get("pageFile");
+        if (!pageFile) {
+          throw new Error("Missing pageFile for Next route tree virtual module.");
+        }
+
+        const entries = await scanNextAppRoutes(root);
+        const entry = entries.find((candidate) => candidate.pageFile === pageFile);
+        if (!entry) {
+          throw new Error(`No Next app route entry found for ${pageFile}.`);
+        }
+
+        const { code, watchFiles } = await generateNextRouteTreeModule(root, entry, entries);
+        for (const file of watchFiles) {
+          this.addWatchFile(file);
+        }
+        return code;
+      }
+
+      if (id === virtualNextRouteManifestId) {
+        const entries = await scanNextAppRoutes(root);
+        for (const entry of entries) {
+          this.addWatchFile(entry.pageFile);
+        }
+
+        return generateNextRouteManifest(entries);
+      }
+    },
+  };
+}
+
+function isInNextAppDir(root: string, file: string) {
+  const dirs = [path.join(root, "app"), path.join(root, "src", "app")];
+  return dirs.some((dir) => {
+    const relative = path.relative(dir, file);
+    return relative && !relative.startsWith("..") && !path.isAbsolute(relative);
+  });
+}
+
+async function scanNextAppRoutes(root: string): Promise<NextRouteManifestBuildEntry[]> {
+  const requireFromProject = createProjectRequire(root);
+  const { findDir } = requireFromProject(
+    "next/dist/lib/find-pages-dir.js",
+  ) as typeof import("next/dist/lib/find-pages-dir.js");
+  const appDir = findDir(root, "app");
+  if (!appDir) return [];
+
+  // Begin copy: Next.js dev app-page route matcher setup
+  // Source: https://github.com/vercel/next.js/blob/4588a7354283f97e2124e3d82f55733ca4eb9373/packages/next/src/server/route-matcher-providers/dev/dev-app-page-route-matcher-provider.ts#L16-L83
+  // Source: https://github.com/vercel/next.js/blob/4588a7354283f97e2124e3d82f55733ca4eb9373/packages/next/src/server/route-matcher-providers/dev/helpers/file-reader/default-file-reader.ts#L12-L42
+  // Adaptation: Vitest asks the provider for matchers directly instead of
+  // running the full Next dev server route matcher manager.
+  const { DevAppPageRouteMatcherProvider } = requireFromProject(
+    "next/dist/server/route-matcher-providers/dev/dev-app-page-route-matcher-provider.js",
+  ) as typeof import("next/dist/server/route-matcher-providers/dev/dev-app-page-route-matcher-provider.js");
+  const { DefaultFileReader } = requireFromProject(
+    "next/dist/server/route-matcher-providers/dev/helpers/file-reader/default-file-reader.js",
+  ) as typeof import("next/dist/server/route-matcher-providers/dev/helpers/file-reader/default-file-reader.js");
+  const { sortPageObjects } = requireFromProject(
+    "next/dist/shared/lib/router/utils/sortable-routes.js",
+  ) as typeof import("next/dist/shared/lib/router/utils/sortable-routes.js");
+
+  const provider = new DevAppPageRouteMatcherProvider(
+    appDir,
+    ["tsx", "ts", "jsx", "js", "mdx"],
+    new DefaultFileReader({
+      ignorePartFilter: (part: string) => part === "node_modules" || part.startsWith("."),
+    }),
+    false,
+  );
+  const matchers = await provider.matchers();
+  // End copy
+
+  const pageFileByAppPath = new Map<string, string>();
+  const matcherByRoute = new Map<string, (typeof matchers)[number]>();
+
+  for (const matcher of matchers) {
+    pageFileByAppPath.set(matcher.definition.page, matcher.definition.filename);
+    matcherByRoute.set(matcher.definition.pathname, matcher);
+  }
+
+  const entries = Array.from(matcherByRoute, ([route, matcher]) => {
+    const appPath = matcher.definition.appPaths.at(-1) ?? matcher.definition.page;
+    const pageFile = pageFileByAppPath.get(appPath) ?? matcher.definition.filename;
+
+    return {
+      route,
+      appDir,
+      appPath,
+      appPaths: matcher.definition.appPaths,
+      allNormalizedAppPaths: Array.from(matcherByRoute.keys()),
+      pageFile,
+    };
+  });
+
+  return Array.from(sortPageObjects(entries, (entry) => entry.route));
+}
+
+function generateNextRouteManifest(entries: NextRouteManifestBuildEntry[]) {
+  const imports = entries
+    .map((entry, index) => {
+      const params = new URLSearchParams({ pageFile: entry.pageFile });
+      return `import { loaderTree as loaderTree${index} } from ${JSON.stringify(`${virtualNextRouteTreePublicId}?${params}`)};`;
+    })
+    .join("\n");
+
+  const manifest = `[${entries
+    .map(
+      (entry, index) => `{
+        route: ${JSON.stringify(entry.route)},
+        appPath: ${JSON.stringify(entry.appPath)},
+        pageFile: ${JSON.stringify(entry.pageFile)},
+        loaderTree: loaderTree${index},
+      }`,
+    )
+    .join(",")}]`;
+
+  return `${imports}\nexport const nextRouteManifest = ${manifest};\n`;
+}
+
+async function generateNextRouteTreeModule(
+  root: string,
+  entry: NextRouteManifestBuildEntry,
+  entries: NextRouteManifestBuildEntry[],
+) {
+  assertRootLayoutExists(entry);
+
+  const watchFiles = new Set<string>([entry.pageFile]);
+  const loader = createProjectRequire(root)(
+    "next/dist/build/webpack/loaders/next-app-loader/index.js",
+  ) as {
+    default: (this: NextAppLoaderContext) => Promise<string>;
+  };
+  // Begin copy: Next.js next-app-loader option shape
+  // Source: https://github.com/vercel/next.js/blob/4588a7354283f97e2124e3d82f55733ca4eb9373/packages/next/src/build/entries.ts#L584-L615
+  // Source: https://github.com/vercel/next.js/blob/4588a7354283f97e2124e3d82f55733ca4eb9373/packages/next/src/build/webpack/loaders/next-app-loader/index.ts#L92-L127
+  // Adaptation: Vite invokes the loader in-process to get the real loader tree
+  // while replacing webpack's module graph with Vite virtual modules.
+  const options = {
+    name: `app${entry.appPath}`,
+    page: entry.appPath,
+    pagePath: `private-next-app-dir${entry.appPath}`,
+    appDir: entry.appDir,
+    appPaths: entry.appPaths,
+    allNormalizedAppPaths: entry.allNormalizedAppPaths,
+    preferredRegion: undefined,
+    pageExtensions: ["tsx", "ts", "jsx", "js", "mdx"],
+    assetPrefix: "",
+    rootDir: root,
+    tsconfigPath: path.join(root, "tsconfig.json"),
+    isDev: false as const,
+    basePath: "",
+    nextConfigOutput: undefined,
+    middlewareConfig: Buffer.from(JSON.stringify({ matchers: [] })).toString("base64"),
+    isGlobalNotFoundEnabled: undefined,
+  };
+  // End copy
+  const context: NextAppLoaderContext = {
+    getOptions: () => options,
+    _module: { buildInfo: {} },
+    _compiler: { context: root },
+    _compilation: undefined,
+    addMissingDependency(file) {
+      watchFiles.add(file);
+    },
+  };
+
+  const generated = await loader.default.call(context);
+  return {
+    code: extractNextLoaderTreeModule(generated),
+    watchFiles: [...watchFiles],
+  };
+}
+
+type NextAppLoaderContext = {
+  getOptions(): {
+    name: string;
+    page: string;
+    pagePath: string;
+    appDir: string;
+    appPaths: readonly string[];
+    allNormalizedAppPaths: readonly string[];
+    preferredRegion: undefined;
+    pageExtensions: string[];
+    assetPrefix: string;
+    rootDir: string;
+    tsconfigPath: string;
+    isDev: false;
+    basePath: string;
+    nextConfigOutput: undefined;
+    middlewareConfig: string;
+    isGlobalNotFoundEnabled: undefined;
+  };
+  _module: { buildInfo: Record<string, unknown> };
+  _compiler: { context: string };
+  _compilation: undefined;
+  addMissingDependency(file: string): void;
+};
+
+function assertRootLayoutExists(entry: NextRouteManifestBuildEntry) {
+  let currentDir = path.dirname(entry.pageFile);
+  while (currentDir.startsWith(entry.appDir)) {
+    if (findAppFile(currentDir, "layout")) return;
+    if (currentDir === entry.appDir) break;
+    currentDir = path.dirname(currentDir);
+  }
+
+  throw new Error(
+    `Cannot render Next route ${entry.route} because no root layout was found above ${entry.pageFile}. Add an app/layout file or render a React node directly.`,
+  );
+}
+
+function findAppFile(dir: string, basename: string) {
+  const extensions = [".tsx", ".ts", ".jsx", ".js", ".mdx"];
+  return extensions.find((ext) => fs.existsSync(path.join(dir, `${basename}${ext}`)));
+}
+
+function extractNextLoaderTreeModule(generated: string) {
+  const treeStart = generated.indexOf("const tree =");
+  const treeEnd = generated.indexOf("const __next_app_require__", treeStart);
+  if (treeStart < 0 || treeEnd < 0) {
+    throw new Error("Could not extract loader tree from Next app loader output.");
+  }
+
+  const treeCode = generated
+    .slice(treeStart, treeEnd)
+    .replace("const tree =", "export const loaderTree =");
+  const referencedLoaders = new Set(
+    Array.from(treeCode.matchAll(/\b([A-Za-z_$][\w$]*)\b/g), (match) => match[1]),
+  );
+  const moduleLoaders = generated
+    .slice(0, treeStart)
+    .split("\n")
+    .filter((line) => {
+      const match = line.match(/^const (\w+) = \(\) => import\(/);
+      return match && referencedLoaders.has(match[1]!);
+    })
+    .join("\n");
+
+  return rewriteNextAppLoaderImports(`${moduleLoaders}\n${treeCode}`);
+}
+
+function rewriteNextAppLoaderImports(code: string) {
+  return code
+    .replace(
+      /import\((\/\*[\s\S]*?\*\/\s*)?(['"])([^'"]+)\2\)/g,
+      (_match, comment: string | undefined, _quote: string, source: string) =>
+        `import(${comment ?? ""}${JSON.stringify(toViteImportSource(source))})`,
+    )
+    .replace(
+      /(from\s+)(['"])([^'"]+)\2/g,
+      (_match, prefix: string, _quote: string, source: string) =>
+        `${prefix}${JSON.stringify(toViteImportSource(source))}`,
+    )
+    .replace(
+      /(import\s+)(['"])([^'"]+)\2/g,
+      (_match, prefix: string, _quote: string, source: string) =>
+        `${prefix}${JSON.stringify(toViteImportSource(source))}`,
+    );
+}
+
+function toViteImportSource(source: string) {
+  if (!path.isAbsolute(source)) return source;
+  const [file] = source.split("?");
+  if (file && !fs.existsSync(file)) return virtualNextRouteEmptyModulePublicId;
+
+  return `/@fs/${normalizePath(source).replace(/^\//, "")}`;
+}
